use handling fee as a percentage

+++ lib/spree/active_shipping_configuration.rb
@@ -17,7 +17,7 @@ class Spree::ActiveShippingConfiguration < Spree::Preferences::Configuration
   preference :units, :string, :default => "imperial"
   preference :unit_multiplier, :integer, :default => 16 # 16 oz./lb - assumes variant weights are in lbs
   preference :default_weight, :integer, :default => 0 # 16 oz./lb - assumes variant weights are in lbs
-  preference :handling_fee, :integer
+  preference :handling_fee, :decimal
   preference :max_weight_per_package, :integer, :default => 0 # 0 means no limit
 
   preference :test_mode, :boolean, :default => false
